Adjust texture priorities in preparation of impl-thread eviction of only some textures
This came from the already-reviewed WebKit bug
https://bugs.webkit.org/show_bug.cgi?id=95758

I forgot to cq? that bug, so it didn't land.

Review URL: https://chromiumcodereview.appspot.com/10917281

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@157075 0039d316-1c4b-4281-b951-d872f2087c98

+++ cc/CCPrioritizedTexture.h
@@ -109,7 +109,7 @@ private:
     GC3Denum m_format;
     size_t m_bytes;
 
-    size_t m_priority;
+    int m_priority;
     bool m_isAbovePriorityCutoff;
     bool m_isSelfManaged;
